[Bugfix][LoRA] Fix dtype mismatch in XPU and CPU punica wrappers
Use input tensor dtype (x.dtype) instead of hardcoded torch.float32
for intermediate buffers in add_lora_linear and add_lora_logits.
This matches the GPU implementation and prevents dtype mismatches
when using bfloat16 or other non-float32 dtypes.

Added test_punica_wrapper_dtype.py to verify buffer dtype matches
input dtype for float16, bfloat16, and float32.

+++ vllm/lora/punica_wrapper/punica_cpu.py
@@ -300,10 +300,8 @@ class PunicaWrapperCPU(PunicaWrapperBase):
 
         if buffer is None:
             r = lora_b_stacked[0].size(-1)
-            # We set the buffer to be float32 by default, consistent with the
-            # triton op
             buffer = tuple(
-                torch.zeros((x.size(0), r), dtype=torch.float32, device=x.device)
+                torch.zeros((x.size(0), r), dtype=x.dtype, device=x.device)
                 for _ in range(len(output_slices))
             )
         self.add_shrink(buffer, x, lora_a_stacked, scale, **kwargs)
@@ -342,9 +340,7 @@ class PunicaWrapperCPU(PunicaWrapperBase):
         x = x.view(-1, x.shape[-1])
         r = lora_b_stacked.size(-1)
         if buffer is None:
-            # We set the buffer to be float32 by default, consistent with the
-            # triton op
-            buffer = torch.zeros((x.size(0), r), dtype=torch.float32, device=x.device)
+            buffer = torch.zeros((x.size(0), r), dtype=x.dtype, device=x.device)
         # LogitsProcessorWithLoRA always using bgmv.
         bgmv_shrink(x, lora_a_stacked, buffer, self.sampler_indices, scale)
         bgmv_expand(buffer, lora_b_stacked, y, self.sampler_indices, add_inputs=True)
